MENUS-28 : factor common code of MainTreeMenuInclude and MainTreeMenuIncludeAllPages into an abstract base class
